# Optimize native histograms