--- a/execution/scan/functions.go
+++ b/execution/scan/functions.go
@@ -294,7 +294,6 @@ func extrapolatedRate(samples []Sample, isCounter, isRate bool, stepTime int64, 
 		resultValue *= factor
 	} else {
 		resultHistogram.Mul(factor)
-
 	}
 
 	return resultValue, resultHistogram