This commit optimizes queries against native histograms by reusing histogram objects for windowing functions.
    
The same optimization was applied in Prometheus in https://github.com/prometheus/prometheus/pull/13276 and https://github.com/prometheus/prometheus/pull/13289.

```
NativeHistograms/selector-8               128ms ± 2%     127ms ± 2%     ~     (p=0.841 n=5+5)
NativeHistograms/sum-8                    174ms ±13%     164ms ± 2%     ~     (p=0.421 n=5+5)
NativeHistograms/rate-8                   359ms ± 3%     278ms ± 1%  -22.42%  (p=0.008 n=5+5)
NativeHistograms/sum_rate-8               341ms ± 5%     265ms ± 5%  -22.28%  (p=0.008 n=5+5)
NativeHistograms/histogram_sum-8          363ms ± 3%     357ms ± 1%   -1.63%  (p=0.016 n=5+5)
NativeHistograms/histogram_count-8        370ms ± 3%     361ms ± 3%     ~     (p=0.056 n=5+5)
NativeHistograms/histogram_quantile-8     174ms ± 1%     174ms ± 3%     ~     (p=0.548 n=5+5)

name                                   old alloc/op   new alloc/op   delta
NativeHistograms/selector-8               434MB ± 0%     434MB ± 0%     ~     (p=1.000 n=5+5)
NativeHistograms/sum-8                    416MB ± 0%     416MB ± 0%     ~     (p=0.310 n=5+5)
NativeHistograms/rate-8                  1.18GB ± 0%    0.87GB ± 0%  -26.15%  (p=0.008 n=5+5)
NativeHistograms/sum_rate-8              1.16GB ± 0%    0.85GB ± 0%  -26.58%  (p=0.008 n=5+5)
NativeHistograms/histogram_sum-8          436MB ± 0%     436MB ± 0%     ~     (p=0.841 n=5+5)
NativeHistograms/histogram_count-8        445MB ± 2%     440MB ± 2%     ~     (p=0.841 n=5+5)
NativeHistograms/histogram_quantile-8     431MB ± 0%     431MB ± 0%     ~     (p=0.151 n=5+5)

name                                   old allocs/op  new allocs/op  delta
NativeHistograms/selector-8               5.19M ± 0%     5.19M ± 0%     ~     (p=0.548 n=5+5)
NativeHistograms/sum-8                    5.19M ± 0%     5.19M ± 0%     ~     (p=1.000 n=5+5)
NativeHistograms/rate-8                   21.1M ± 0%     14.7M ± 0%  -30.39%  (p=0.008 n=5+5)
NativeHistograms/sum_rate-8               21.1M ± 0%     14.7M ± 0%  -30.40%  (p=0.008 n=5+5)
NativeHistograms/histogram_sum-8          5.20M ± 0%     5.20M ± 0%     ~     (p=0.548 n=5+5)
NativeHistograms/histogram_count-8        5.24M ± 1%     5.22M ± 1%     ~     (p=0.548 n=5+5)
NativeHistograms/histogram_quantile-8     5.26M ± 0%     5.26M ± 0%     ~     (p=0.111 n=5+5)
```