@fpetkovski 2023-12-29
Yes, but we also had to change the matrix selector to reuse histogram objects. Some small tweaks will be needed after  https://github.com/prometheus/prometheus/pull/13340 is merged.